Feat | Extract cache aux libs to file
